void private messages

+++ cogs/moderator.py
@@ -23,25 +23,26 @@ class Events(Cog):
 
     @Cog.listener(name="on_message")
     async def restricted_messages(self, message: Message):
-        channels_conf = self.config \
-            .find("Guilds") \
-            .find(f"Guild-{message.guild.id}") \
-            .find("Channels")
-        try:
-            restricted_channels = {channels_conf.find("BOT_COMMANDS_CHANNEL").text, channels_conf.find("HELP").text}
-            if message.guild and len(message.author.roles) == 1 and message.channel.id in restricted_channels:
-                match = re.match(self.link, message.content)
-                if match:
-                    await message.delete()
-                    await message.channel.send(f"<@{message.author.id}>\n"
-                                               f"Non-verified members are not allowed to post links to this channel.")
-                    logger.info(f"Message >>{message.clean_content}<< in channel {message.channel}"
-                                f"from {message.author.display_name}({message.author.id}) "
-                                f"deleted according due restrictions.")
-        except Forbidden:
-            logger.warning("Failed to fulfill restriction.")
-        except HTTPException:
-            pass
+        if message.guild:
+            channels_conf = self.config \
+                .find("Guilds") \
+                .find(f"Guild-{message.guild.id}") \
+                .find("Channels")
+            try:
+                restricted_channels = {channels_conf.find("BOT_COMMANDS_CHANNEL").text, channels_conf.find("HELP").text}
+                if message.guild and len(message.author.roles) == 1 and message.channel.id in restricted_channels:
+                    match = re.match(self.link, message.content)
+                    if match:
+                        await message.delete()
+                        await message.channel.send(f"<@{message.author.id}>\n"
+                                                   f"Non-verified members are not allowed to post links to this channel.")
+                        logger.info(f"Message >>{message.clean_content}<< in channel {message.channel}"
+                                    f"from {message.author.display_name}({message.author.id}) "
+                                    f"deleted according due restrictions.")
+            except Forbidden:
+                logger.warning("Failed to fulfill restriction.")
+            except HTTPException:
+                pass
 
 
 class Moderator(Cog):
